+++ nringthymeleaf/src/main/java/com/katamlek/nringthymeleaf/domain/PaymentNote.java
@@ -5,6 +5,7 @@ import lombok.Data;
 import javax.persistence.Entity;
 import javax.persistence.ManyToOne;
 import javax.validation.constraints.NotNull;
+import java.math.BigDecimal;
 
 @Entity
 @Data
@@ -12,7 +13,7 @@ public class PaymentNote extends Note {
     private String paymentMethod;
 
     @NotNull
-    private Long paymentAmount;
+    private BigDecimal paymentAmount;
 
     @ManyToOne
     Booking booking;
